Tweak syntax & tidy up a bit
As you can probably tell, I'm bored.

Tweaked the comments, because they were going pretty much to the other
end of the monitor, which is a bit nuts for other people reading your
code. :P

Changed use of `netcat` to `ping`, because it's more portable and
simpler. If there are any issues or you for some reason _don't_ have
`ping`, you'll of course need to change it back.

Instead of separately using `$?`, you can just directly test for status
by putting the command into the `if` statement as the condition,
because remember, that's basically what an `if` statement, `while`
loop, and `until` loop is doing; they're checking the status of a
command.